Models and Views Folders - Move the 2 windows out of Views Dir
