+++ App/MainWindow.xaml.cs
@@ -1,20 +1,20 @@
-﻿using Hurl.Browser;
+﻿using Hurl.Models;
 using System;
 using System.Diagnostics;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 
-namespace Hurl.Views
+namespace Hurl
 {
     /// <summary>
-    /// Interaction logic for SelectionWindow.xaml
+    /// Interaction logic for MainWindow.xaml
     /// </summary>
-    public partial class SelectionWindow : Window
+    public partial class MainWindow : Window
     {
         private string OpenedLink = null;
 
-        public SelectionWindow(string[] x)
+        public MainWindow(string[] x)
         {
             InitializeComponent();
