+++ App/SettingsWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="Hurl.Views.SettingsWindow"
+﻿<Window x:Class="Hurl.SettingsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:uc="clr-namespace:Hurl.Controls" xmlns:local="clr-namespace:Hurl.Views"
+        xmlns:uc="clr-namespace:Hurl.Controls" xmlns:local="clr-namespace:Hurl"
         mc:Ignorable="d" ResizeMode="NoResize" WindowStartupLocation="CenterScreen"
         Title="Hurl Settings" Width="600" Height="400">
     <TabControl Background="Azure">
