--- App/Views/SelectionWindow.xaml
@@ -1,40 +1,0 @@
-﻿<Window x:Class="Hurl.Views.SelectionWindow" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-        xmlns:local="clr-namespace:Hurl.Views" mc:Ignorable="d" 
-        ResizeMode="NoResize" Topmost="True" WindowStyle="None" WindowStartupLocation="CenterScreen" 
-        Title="Hurl - Select a Browser" Width="320" Height="210" AllowsTransparency="True" Background="{x:Null}"  
-        Loaded="Window_Loaded" Deactivated="Window_Deactivated" KeyDown="Window_Esc">
-    <Window.Clip>
-        <RectangleGeometry Rect="0,0,320,210" RadiusX="15" RadiusY="15" />
-    </Window.Clip>
-    <Window.Resources>
-        <ContextMenu x:Key="extraOptions">
-            <MenuItem Header="Incognito" Click="Incognito_Click" />
-            <Separator />
-            <MenuItem Header="Profile" />
-        </ContextMenu>
-
-        <Style x:Key="BrowserBtnStyle" TargetType="Button">
-            <Setter Property="Foreground" Value="Black"/>
-            <Setter Property="Background" Value="White" />
-            <Setter Property="BorderBrush" Value="Black" />
-            <Setter Property="BorderThickness" Value="1" />
-        </Style>
-    </Window.Resources>
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="25" />
-            <RowDefinition Height="1*" />
-            <RowDefinition Height="25" />
-        </Grid.RowDefinitions>
-        <Grid>
-            <TextBlock x:Name="header" Padding="15,3,0,0" FontSize="14" Background="White" Foreground="#FFE02B47" FontFamily="Verdana">Select a Browser</TextBlock>
-            <Button HorizontalAlignment="Right" VerticalAlignment="Center" Margin="0,0,18,0" FontSize="10" Click="OpenSettings" Padding="5,2,5,2" BorderBrush="Black" Background="White">Settings</Button>
-            <Button HorizontalAlignment="Right" VerticalAlignment="Center" Margin="0,0,72,0" FontSize="10" Click="LinkCopyBtn" Padding="5,2,5,2" BorderBrush="Black" Background="White">Copy Link</Button>
-        </Grid>
-        <StackPanel Background="#FFE02B47" x:Name="stacky" Grid.Row="1" />
-        <TextBlock x:Name="linkpreview" Padding="2,6,2,3" TextAlignment="Center" FontSize="10" Grid.Row="2" Background="White" FontFamily="Calibri">No Link Opened</TextBlock>
-    </Grid>
-</Window>
